requirements.txt
nltk==3.7
pdf2image==1.16.0
Pillow==9.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
spacy==3.3.1
